Revert "Extract blog contents to Engine"
This reverts commit ef3a6a8d06389c2c21a9f1e46e450d13f4cad544.

Conflicts:
	daimon_news_blog/app/views/layouts/daimon_news_blog/application.html.erb
