+++ app/views/site_posts/_long_collection.html.erb
@@ -1,7 +1,7 @@
 <ul>
   <% posts.each do |post| %>
     <li>
-      <%#= link_to post.title, post_path(post.id) %>
+      <%= link_to post.title, post_path(post.id) %>
       <p><%= post.body %></p>
     </li>
   <% end %>
