+++ app/views/site_posts/_short_collection.html.erb
@@ -1,7 +1,7 @@
 <ol>
   <% posts.each do |post| %>
     <li>
-      <%#= link_to post.title, post_path(post.id) %>
+      <%= link_to post.title, post_path(post.id) %>
     </li>
   <% end %>
 </ol>
