+++ app/views/site_posts/index.html.erb
