+++ app/views/site_posts/index.json.rabl
