+++ app/views/site_posts/show.json.rabl
