--- daimon_news_blog/app/models/daimon_news_blog/post.rb
@@ -1,4 +1,0 @@
-module DaimonNewsBlog
-  class Post < DaimonNewsBlog.post_class.constantize
-  end
-end
