--- daimon_news_blog/app/models/daimon_news_blog/site.rb
@@ -1,4 +1,0 @@
-module DaimonNewsBlog
-  class Site < DaimonNewsBlog.site_class.constantize
-  end
-end
