+++ daimon_news_blog/lib/daimon_news_blog.rb
@@ -1,6 +1,4 @@
 require "daimon_news_blog/engine"
 
 module DaimonNewsBlog
-  mattr_accessor :site_class
-  mattr_accessor :post_class
 end
